Se cambian de float a uint32_t en funcion libreria lcd para mostrar por pantalla valores

+++ freeRTOSHuerta/Core/Inc/bsp.h
@@ -27,8 +27,8 @@
 
 void BSP_Init(void);
 void BSP_Delay(uint16_t);
-void BSP_LCD_Temperature(float);
-void BSP_LCD_Humidity(float);
+void BSP_LCD_Temperature(uint32_t);
+void BSP_LCD_Humidity(uint32_t);
 void BSP_LCD_SoilHumidity(uint32_t);
 uint32_t BSP_Get_percentageHS(uint32_t);
 void BSP_Detect_Movement();
